requirement.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-postgres-utils==0.1.2
djangorestframework==3.14.0
psycopg2==2.9.4
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.5
waitress==2.1.2
whitenoise==6.2.0
